When deleting characters, continue completing
When asyncomplete doesn't refresh a source, it doesn't also call core_complete using previously cached values. This commit allows asyncomplete to return cached values when a source isn't refreshed. This could probably be improved - for instance, there may be a way to complete cached values for ALL sources that aren't refreshed instead of only when there are NO refreshed sources