Allow custom filter for sources
This patch allows sources to have its own candidate-filtering function.
If a source has the 'filter' key, it is used instead of the internal
filter. Its value should be a funcref which takes two arguments;
prefix and candidates. The `prefix` is the text before the cursor and
`candidates` is the list of completion candidates.