Only warn about missing timers functionality once
plugin/asyncomplete.vim sets up an autocmd to invoke
asyncomplete#enable_for_buffer() for every BufEnter. If we don't clear
this autocmd, then it will try to call that function, which will try to
source the autoload file, which outputs the error again.

Clearing that autocmd group stops the cycle and limits the number of
warnings to one.